Move some project routes out to traits
