+++ app-backend/api/src/main/scala/project/Routes.scala
@@ -21,9 +21,8 @@ import com.rasterfoundry.akkautil.{
 }
 import com.rasterfoundry.database._
 import com.rasterfoundry.database.filter.Filterables._
-import com.rasterfoundry.common.datamodel.GeoJsonCodec._
 import com.rasterfoundry.common.datamodel.{Annotation, _}
-import com.lonelyplanet.akka.http.extensions.{PageRequest, PaginationDirectives}
+import com.lonelyplanet.akka.http.extensions.PaginationDirectives
 import com.typesafe.scalalogging.LazyLogging
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import doobie._
@@ -32,17 +31,10 @@ import doobie.util.transactor.Transactor
 import geotrellis.shapefile.ShapeFileReader
 import io.circe.generic.JsonCodec
 
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.Implicits.global
 import scala.util.Success
 
 @JsonCodec
 final case class BulkAcceptParams(sceneIds: List[UUID])
-
-@JsonCodec
-final case class AnnotationFeatureCollectionCreate(
-    features: Seq[Annotation.GeoJSONFeatureCreate]
-)
 
 trait ProjectRoutes
     extends Authentication
@@ -56,60 +48,12 @@ trait ProjectRoutes
     with AWSBatch
     with UserErrorHandler
     with RollbarNotifier
-    with LazyLogging {
+    with LazyLogging
+    with ProjectAnnotationRoutes
+    with ProjectLayerRoutes
+    with ProjectLayerAnnotationRoutes {
 
   val xa: Transactor[IO]
-
-  val BULK_OPERATION_MAX_LIMIT = 100
-
-  private def projectIsPublic(projectId: UUID): Directive0 = {
-    authorizeAsync {
-      ProjectDao
-        .unsafeGetProjectById(projectId)
-        .transact(xa)
-        .unsafeToFuture map {
-        _.visibility == Visibility.Public
-      }
-    }
-  }
-
-  private def projectAuthFromMapTokenO(mapTokenO: Option[UUID],
-                                       projectId: UUID): Directive0 = {
-    mapTokenO map { mapToken =>
-      authorizeAsync {
-        MapTokenDao
-          .checkProject(projectId)(mapToken)
-          .transact(xa)
-          .map({
-            case Some(_) => true
-            case _       => false
-          })
-          .unsafeToFuture
-      }
-    } getOrElse { reject(AuthorizationFailedRejection) }
-  }
-
-  private def projectAuthFromTokenO(tokenO: Option[String],
-                                    projectId: UUID): Directive0 = {
-    authorizeAsync {
-      tokenO map { token =>
-        verifyJWT(token) traverse {
-          case (_, jwtClaims) =>
-            val userId = jwtClaims.getStringClaim("sub")
-            for {
-              user <- UserDao.unsafeGetUserById(userId)
-              projectAuth <- ProjectDao.authorized(user,
-                                                   ObjectType.Project,
-                                                   projectId,
-                                                   ActionType.View)
-            } yield projectAuth
-        } map {
-          case Right(result) => result
-          case Left(_)       => false
-        } transact (xa) unsafeToFuture
-      } getOrElse { Future.successful(false) }
-    }
-  }
 
   val projectRoutes: Route = handleExceptions(userExceptionHandler) {
     pathEndOrSingleSlash {
@@ -662,266 +606,6 @@ trait ProjectRoutes
     } {
       complete {
         AnnotationDao.listProjectLabels(projectId).transact(xa).unsafeToFuture
-      }
-    }
-  }
-
-  def listAnnotationGroups(projectId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authorized(user, ObjectType.Project, projectId, ActionType.View)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      complete {
-        AnnotationGroupDao
-          .listForProject(projectId)
-          .transact(xa)
-          .unsafeToFuture
-      }
-    }
-  }
-
-  def createAnnotationGroup(projectId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      entity(as[AnnotationGroup.Create]) { agCreate =>
-        complete {
-          AnnotationGroupDao
-            .createAnnotationGroup(projectId, agCreate, user)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-    }
-  }
-
-  def getAnnotationGroup(projectId: UUID, agId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .getAnnotationGroup(projectId, agId)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-  }
-
-  def getAnnotationGroupSummary(projectId: UUID,
-                                annotationGroupId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .getAnnotationGroupSummary(projectId, annotationGroupId)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-  }
-
-  def updateAnnotationGroup(projectId: UUID, agId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[AnnotationGroup]) { annotationGroup =>
-          complete {
-            AnnotationGroupDao
-              .updateAnnotationGroup(projectId, annotationGroup, agId, user)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-  }
-
-  def deleteAnnotationGroup(projectId: UUID, agId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .deleteAnnotationGroup(projectId, agId)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-  }
-
-  def listAnnotations(projectId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authorized(user, ObjectType.Project, projectId, ActionType.View)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      (withPagination & annotationQueryParams) {
-        (page: PageRequest, queryParams: AnnotationQueryParameters) =>
-          complete {
-            AnnotationDao
-              .listByLayer(projectId, page, queryParams)
-              .transact(xa)
-              .unsafeToFuture
-              .map { p =>
-                {
-                  fromPaginatedResponseToGeoJson[Annotation,
-                                                 Annotation.GeoJSON](p)
-                }
-              }
-          }
-      }
-    }
-  }
-
-  def createAnnotation(projectId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      entity(as[AnnotationFeatureCollectionCreate]) { fc =>
-        val annotationsCreate = fc.features map { _.toAnnotationCreate }
-        onSuccess(
-          AnnotationDao
-            .insertAnnotations(annotationsCreate.toList, projectId, user)
-            .transact(xa)
-            .unsafeToFuture
-            .map { annotations: List[Annotation] =>
-              fromSeqToFeatureCollection[Annotation, Annotation.GeoJSON](
-                annotations)
-            }
-        ) { createdAnnotation =>
-          complete((StatusCodes.Created, createdAnnotation))
-        }
-      }
-    }
-  }
-
-  def exportAnnotationShapefile(projectId: UUID): Route = authenticate { user =>
-    (annotationExportQueryParameters) { annotationExportQP =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          AnnotationDao
-            .listForProjectExport(projectId, annotationExportQP)
-            .transact(xa)
-            .unsafeToFuture) {
-          case annotations @ (annotation: List[Annotation]) => {
-            complete(
-              AnnotationShapefileService
-                .getAnnotationShapefileDownloadUrl(annotations, user)
-            )
-          }
-          case _ =>
-            complete(
-              throw new Exception(
-                "Annotations do not exist or are not accessible by this user"))
-        }
-      }
-    }
-  }
-
-  def getAnnotation(projectId: UUID, annotationId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        rejectEmptyResponse {
-          complete {
-            AnnotationDao
-              .getAnnotationById(projectId, annotationId)
-              .transact(xa)
-              .unsafeToFuture
-              .map {
-                _ map { _.toGeoJSONFeature }
-              }
-          }
-        }
-      }
-  }
-
-  def updateAnnotation(projectId: UUID, annotationId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[Annotation.GeoJSON]) {
-          updatedAnnotation: Annotation.GeoJSON =>
-            onSuccess(AnnotationDao
-              .updateAnnotation(projectId, updatedAnnotation.toAnnotation, user)
-              .transact(xa)
-              .unsafeToFuture) { count =>
-              completeSingleOrNotFound(count)
-            }
-        }
-      }
-    }
-
-  def deleteAnnotation(projectId: UUID, annotationId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          AnnotationDao
-            .deleteById(projectId, annotationId)
-            .transact(xa)
-            .unsafeToFuture) {
-          completeSingleOrNotFound
-        }
-      }
-    }
-
-  def deleteProjectAnnotations(projectId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      onSuccess(
-        AnnotationDao
-          .deleteByProjectLayer(projectId)
-          .transact(xa)
-          .unsafeToFuture) {
-        completeSomeOrNotFound
       }
     }
   }
@@ -1493,567 +1177,4 @@ trait ProjectRoutes
       .toList
   }
 
-  def createProjectLayer(projectId: UUID): Route = authenticate { user =>
-    entity(as[ProjectLayer.Create]) { newProjectLayer =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          ProjectLayerDao
-            .insertProjectLayer(newProjectLayer.toProjectLayer)
-            .transact(xa)
-            .unsafeToFuture) { projectLayer =>
-          complete(StatusCodes.Created, projectLayer)
-        }
-      }
-    }
-  }
-
-  def listProjectLayers(projectId: UUID): Route = extractTokenHeader { tokenO =>
-    extractMapTokenParam { mapTokenO =>
-      (projectAuthFromMapTokenO(mapTokenO, projectId) |
-        projectAuthFromTokenO(tokenO, projectId) | projectIsPublic(projectId)) {
-        (withPagination) { (page) =>
-          complete {
-            ProjectLayerDao
-              .listProjectLayersForProject(page, projectId)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-
-    }
-  }
-
-  def getProjectLayer(projectId: UUID, layerId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        rejectEmptyResponse {
-          complete {
-            ProjectLayerDao
-              .getProjectLayer(projectId, layerId)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-  }
-
-  def updateProjectLayer(projectId: UUID, layerId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[ProjectLayer]) { updatedProjectLayer =>
-          onSuccess(
-            ProjectLayerDao
-              .updateProjectLayer(updatedProjectLayer, layerId)
-              .transact(xa)
-              .unsafeToFuture) {
-            completeSingleOrNotFound
-          }
-        }
-      }
-  }
-
-  def deleteProjectLayer(projectId: UUID, layerId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        rejectEmptyResponse {
-          complete {
-            ProjectLayerDao
-              .deleteProjectLayer(layerId)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-  }
-
-  def getProjectLayerMosaicDefinition(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        rejectEmptyResponse {
-          complete {
-            SceneToLayerDao
-              .getMosaicDefinition(layerId)
-              .compile
-              .to[List]
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-    }
-
-  def getProjectLayerSceneColorCorrectParams(projectId: UUID,
-                                             layerId: UUID,
-                                             sceneId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          SceneToLayerDao
-            .getColorCorrectParams(layerId, sceneId)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-    }
-
-  def setProjectLayerSceneColorCorrectParams(projectId: UUID,
-                                             layerId: UUID,
-                                             sceneId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[ColorCorrect.Params]) { ccParams =>
-          onSuccess(
-            SceneToLayerDao
-              .setColorCorrectParams(layerId, sceneId, ccParams)
-              .transact(xa)
-              .unsafeToFuture) { stl =>
-            complete(StatusCodes.NoContent)
-          }
-        }
-      }
-    }
-
-  def setProjectLayerScenesColorCorrectParams(projectId: UUID,
-                                              layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[BatchParams]) { params =>
-          onSuccess(
-            SceneToLayerDao
-              .setColorCorrectParamsBatch(layerId, params)
-              .transact(xa)
-              .unsafeToFuture
-          ) { scenesToLayer =>
-            complete(StatusCodes.NoContent)
-          }
-        }
-      }
-    }
-
-  def setProjectLayerSceneOrder(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Edit)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[Seq[UUID]]) { sceneIds =>
-          if (sceneIds.length > BULK_OPERATION_MAX_LIMIT) {
-            complete(StatusCodes.RequestEntityTooLarge)
-          }
-
-          onSuccess(
-            SceneToLayerDao
-              .setManualOrder(layerId, sceneIds)
-              .transact(xa)
-              .unsafeToFuture
-          ) { updatedOrder =>
-            complete(StatusCodes.NoContent)
-          }
-        }
-      }
-    }
-
-  def listLayerScenes(projectId: UUID, layerId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        (withPagination & projectSceneQueryParameters) { (page, sceneParams) =>
-          complete {
-            ProjectLayerScenesDao
-              .listLayerScenes(layerId, page, sceneParams)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-  }
-
-  def listLayerDatasources(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      (projectQueryParameters) { projectQueryParams =>
-        authorizeAsync {
-          val authorized = for {
-            authProject <- ProjectDao.authorized(user,
-                                                 ObjectType.Project,
-                                                 projectId,
-                                                 ActionType.View)
-            authResult <- (authProject, projectQueryParams.analysisId) match {
-              case (false, Some(analysisId: UUID)) =>
-                ToolRunDao
-                  .authorizeReferencedProject(user, analysisId, projectId)
-              case (_, _) => authProject.pure[ConnectionIO]
-            }
-          } yield authResult
-          authorized.transact(xa).unsafeToFuture
-        } {
-          complete {
-            ProjectLayerDatasourcesDao
-              .listProjectLayerDatasources(layerId)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-    }
-
-  def listLayerLabels(projectId: UUID, layerId: UUID): Route = authenticate {
-    user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationDao
-            .listProjectLabels(projectId, Some(layerId))
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-  }
-
-  def listLayerAnnotations(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        (withPagination & annotationQueryParams) {
-          (page: PageRequest, queryParams: AnnotationQueryParameters) =>
-            complete {
-              AnnotationDao
-                .listByLayer(projectId, page, queryParams, Some(layerId))
-                .transact(xa)
-                .unsafeToFuture
-                .map { p =>
-                  {
-                    fromPaginatedResponseToGeoJson[Annotation,
-                                                   Annotation.GeoJSON](p)
-                  }
-                }
-            }
-        }
-      }
-    }
-
-  def createLayerAnnotation(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[AnnotationFeatureCollectionCreate]) { fc =>
-          val annotationsCreate = fc.features map { _.toAnnotationCreate }
-          onSuccess(
-            AnnotationDao
-              .insertAnnotations(annotationsCreate.toList,
-                                 projectId,
-                                 user,
-                                 Some(layerId))
-              .transact(xa)
-              .unsafeToFuture
-              .map { annotations: List[Annotation] =>
-                fromSeqToFeatureCollection[Annotation, Annotation.GeoJSON](
-                  annotations)
-              }
-          ) { createdAnnotation =>
-            complete((StatusCodes.Created, createdAnnotation))
-          }
-        }
-      }
-    }
-
-  def deleteLayerAnnotations(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          AnnotationDao
-            .deleteByProjectLayer(projectId, Some(layerId))
-            .transact(xa)
-            .unsafeToFuture) {
-          completeSomeOrNotFound
-        }
-      }
-    }
-
-  def getLayerAnnotation(projectId: UUID,
-                         annotationId: UUID,
-                         layerId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      rejectEmptyResponse {
-        complete {
-          AnnotationDao
-            .getAnnotationById(projectId, annotationId)
-            .transact(xa)
-            .unsafeToFuture
-            .map {
-              _ map { _.toGeoJSONFeature }
-            }
-        }
-      }
-    }
-  }
-  def updateLayerAnnotation(projectId: UUID,
-                            annotationId: UUID,
-                            layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[Annotation.GeoJSON]) {
-          updatedAnnotation: Annotation.GeoJSON =>
-            onSuccess(AnnotationDao
-              .updateAnnotation(projectId, updatedAnnotation.toAnnotation, user)
-              .transact(xa)
-              .unsafeToFuture) { count =>
-              completeSingleOrNotFound(count)
-            }
-        }
-      }
-    }
-
-  def deleteLayerAnnotation(projectId: UUID,
-                            annotationId: UUID,
-                            layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          AnnotationDao
-            .deleteById(projectId, annotationId)
-            .transact(xa)
-            .unsafeToFuture) {
-          completeSingleOrNotFound
-        }
-      }
-    }
-
-  def exportLayerAnnotationShapefile(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        onSuccess(
-          AnnotationDao
-            .listForLayerExport(projectId, layerId)
-            .transact(xa)
-            .unsafeToFuture) {
-          case annotations @ (annotation: List[Annotation]) => {
-            complete(
-              AnnotationShapefileService
-                .getAnnotationShapefileDownloadUrl(annotations, user)
-            )
-          }
-          case _ =>
-            complete(
-              throw new Exception(
-                "Annotations do not exist or are not accessible by this user"))
-        }
-      }
-    }
-
-  def listLayerAnnotationGroups(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .listForProject(projectId, Some(layerId))
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-    }
-
-  def createLayerAnnotationGroup(projectId: UUID, layerId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[AnnotationGroup.Create]) { agCreate =>
-          complete {
-            AnnotationGroupDao
-              .createAnnotationGroup(projectId, agCreate, user, Some(layerId))
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-    }
-
-  def getLayerAnnotationGroup(projectId: UUID,
-                              layerId: UUID,
-                              agId: UUID): Route = authenticate { user =>
-    authorizeAsync {
-      ProjectDao
-        .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-        .transact(xa)
-        .unsafeToFuture
-    } {
-      complete {
-        AnnotationGroupDao
-          .getAnnotationGroup(projectId, agId)
-          .transact(xa)
-          .unsafeToFuture
-      }
-    }
-  }
-
-  def updateLayerAnnotationGroup(projectId: UUID,
-                                 layerId: UUID,
-                                 agId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        entity(as[AnnotationGroup]) { annotationGroup =>
-          complete {
-            AnnotationGroupDao
-              .updateAnnotationGroup(projectId, annotationGroup, agId, user)
-              .transact(xa)
-              .unsafeToFuture
-          }
-        }
-      }
-    }
-
-  def deleteLayerAnnotationGroup(projectId: UUID,
-                                 layerId: UUID,
-                                 agId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.Annotate)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .deleteAnnotationGroup(projectId, agId)
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-    }
-
-  def getLayerAnnotationGroupSummary(projectId: UUID,
-                                     layerId: UUID,
-                                     annotationGroupId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authProjectLayerExist(projectId, layerId, user, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          AnnotationGroupDao
-            .getAnnotationGroupSummary(projectId,
-                                       annotationGroupId,
-                                       Some(layerId))
-            .transact(xa)
-            .unsafeToFuture
-        }
-      }
-    }
-
-  def getProjectLayerSceneCounts(projectId: UUID): Route =
-    authenticate { user =>
-      authorizeAsync {
-        ProjectDao
-          .authorized(user, ObjectType.Project, projectId, ActionType.View)
-          .transact(xa)
-          .unsafeToFuture
-      } {
-        complete {
-          ProjectLayerScenesDao
-            .countLayerScenes(projectId)
-            .transact(xa)
-            .map(Map(_: _*))
-            .unsafeToFuture
-        }
-      }
-    }
 }
